+++ code/src/types/form.ts
@@ -147,6 +147,7 @@ interface JsonLdLiteral {
 }
 
 export interface RegistryTaskOption {
+  id: string;
   contract: string;
   status: string;
   date: string;